update: Product CRUD의 권한 변경 2022.08.17
토큰을 가진 관리자에서 토큰을 가진 사업자만 상품의 CRUD 기능을 사용할 수 있다.

+++ api/routes/product.js
@@ -1,15 +1,17 @@
 const Product = require("../models/Product");
+const Shop = require("../models/Shop");
 const {
   verifyToken,
   verifyTokenAndAuthorization,
   verifyTokenAndAdmin,
+  verifyTokenAndEntrepreneur,
 } = require("./verifyToken");
 
 const router = require("express").Router();
 
 //CREATE
-router.post("/", verifyTokenAndAdmin, async (req, res) => {
-  const newProduct = new Product(req.body);
+router.post("/", verifyTokenAndEntrepreneur, async (req, res) => {
+  const newProduct = new Product({ ...req.body, userId: req.user.id });
 
   try {
     const savedProduct = await newProduct.save();
@@ -20,7 +22,7 @@ router.post("/", verifyTokenAndAdmin, async (req, res) => {
 });
 
 //UPDATE
-router.put("/:id", verifyTokenAndAdmin, async (req, res) => {
+router.put("/:id", verifyTokenAndEntrepreneur, async (req, res) => {
   try {
     const updatedProduct = await Product.findByIdAndUpdate(
       req.params.id,
@@ -36,9 +38,12 @@ router.put("/:id", verifyTokenAndAdmin, async (req, res) => {
 });
 
 //DELETE
-router.delete("/:id", verifyTokenAndAdmin, async (req, res) => {
+router.delete("/:id", verifyTokenAndEntrepreneur, async (req, res) => {
+  const product = await Product.findById(req.params.id);
+  const shop = await Shop.findById(req.params.id);
   try {
-    await Product.findByIdAndDelete(req.params.id);
+    if (req.user.id === product.userId || req.user.id === shop.userId)
+      await Product.findByIdAndDelete(req.params.id);
     res.status(200).json("상품이 삭제되었습니다.");
   } catch (err) {
     res.status(500).json(err);
